c/tests: introduced rebalancing tests fixture
Extracted fixture from nodes decommissioning tests. Rebalancing tests
fixture will be used to test partitions replicas rebalancing after node
addition.
